Commit 11/06/2020

+++ app/src/main/java/com/example/tiesiyasuo/Game.java
@@ -3,11 +3,11 @@ package com.example.tiesiyasuo;
 public class Game{
     String name;
     String console;
-    Integer price;
+    Float price;
     String ID;
     String Descrip;
     static int count = 0;
-    public Game(String name, String console, Integer price, String description){
+    public Game(String name, String console, Float price, String description){
         count++;
         String temp = Integer.toString(count);
         while (temp.length() < 5){
@@ -16,11 +16,11 @@ public class Game{
         this.ID = temp;
         this.name = name;
         this.console = console;
-        this.price = price;
+        this.price = (float) price;
         this.Descrip = description;
     }
 
-    public Game(String name, String console, Integer price){
+    public Game(String name, String console, Float price){
         count++;
         String temp = Integer.toString(count);
         while (temp.length() < 5){
@@ -29,7 +29,7 @@ public class Game{
         this.ID = temp;
         this.name = name;
         this.console = console;
-        this.price = price;
+        this.price = (float) price;
         this.Descrip = "...";
     }
 
@@ -45,7 +45,7 @@ public class Game{
         return "$" + this.price;
     }
 
-    public Integer getPriceInt(){
+    public Float getPriceInt(){
         return this.price;
     }
     public String getID(){
